Default disguise for each InfantryType (#373)
* Implement issue #372

* minor adjustments

* Reduce hook count

* Apply suggestions from code review

* Fix stupid repeating

+++ src/Ext/BuildingType/Body.cpp
@@ -23,7 +23,7 @@ int BuildingTypeExt::GetEnhancedPower(BuildingClass* pBuilding, HouseClass* pHou
 		}
 	}
 
-	return std::round(pBuilding->GetPowerOutput() * fFactor) + nAmount;
+	return static_cast<int>(std::round(pBuilding->GetPowerOutput() * fFactor)) + nAmount;
 }
 
 void BuildingTypeExt::ExtData::Initialize()
